created useLocalStorage hook
