+++ pages/index.tsx
@@ -14,6 +14,7 @@ import ListItemText from '@material-ui/core/ListItemText';
 import { EntryFilePicker } from '../components/EntryFilePicker';
 import { ChunksPicker } from '../components/ChunksPicker';
 import Layout from '../components/Layout';
+import { useLocalStorage } from '../components/customHooks/useLocalStorage';
 
 const useStyles = makeStyles((theme) => ({
   flexNone: { flex: '0 0 auto' },
@@ -184,33 +185,6 @@ function Add() {
       </Layout>
     </Box>
   );
-}
-// Sync state to local storage so that it persists through a page refresh.
-// Usage is similar to useState except we pass in a local storage key so that
-// we can default to that value on page load instead of the specified initial value.
-function useLocalStorage(key: string, initialValue: string[]) {
-  const [storedValue, setStoredValue] = useState(() => {
-    if (typeof window === 'undefined') {
-      return initialValue;
-    }
-    try {
-      const item = window.sessionStorage.getItem(key);
-      return item ? JSON.parse(item) : initialValue;
-    } catch (error) {
-      console.log(error);
-      return initialValue;
-    }
-  });
-  const setValue = (value: any) => {
-    try {
-      const valueToStore = value instanceof Function ? value(storedValue) : value;
-      setStoredValue(valueToStore);
-      window.sessionStorage.setItem(key, JSON.stringify(valueToStore));
-    } catch (error) {
-      console.log(error);
-    }
-  };
-  return [storedValue, setValue];
 }
 
 export default Add;
